% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_kegg.R
\name{parse_kegg_groups}
\alias{parse_kegg_groups}
\title{Parse KEGG group nodes from a KGML file}
\usage{
parse_kegg_groups(kgml)
}
\arguments{
\item{kgml}{A parsed XML document from \code{xml2::read_xml()}, representing a KEGG KGML file.}
}
\value{
A list with two elements:
\describe{
\item{mapping}{A tibble mapping each member ID to its meta_id (group_id or itself).}
\item{edges}{A tibble with pairwise intra-group edges (from, to, type = "group").}
}
}
\description{
Extracts group-type entries from a KEGG KGML file and generates both:
(1) a mapping of group ID to component node IDs (with meta_id),
(2) a set of intra-group edges linking the components as undirected pairs.
}
\examples{
kgml_file <- system.file("extdata", "hsa04210.xml", package = "punKEGGer")
doc <- xml2::read_xml(kgml_file)
group_data <- parse_kegg_groups(doc)
head(group_data$mapping)
head(group_data$edges)
}
